Changed location of kajamtag.h
You'll need to use the latest git version of kajamtag and install with autotools to use
kajamtag now.  Kajammer will properly link against the new libkajamtag.la now, but it's in
/usr/local/lib, which isn't in the default LD_PATH for me.  You may need to add the
directory to $LD_LIBRARY_PATH

I'm still learning autotools, so it won't be installed to /usr/local/lib for long.